backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.43
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.10.1,<3.0.0
email-validator==2.1.1
alembic==1.12.1
python-multipart==0.0.9
pytest==7.4.4
httpx==0.27.2
google-generativeai>=0.8.5
Pillow==10.4.0
Jinja2>=3.1.4,<4.0.0
